Sampling easy and hard validation episodes during training.
Previous code sampled only the 200 easy validation episodes during training which resulted in validation results appearing higher than they actually were. This fix now samples easy and hard episodes so metrics are not inflated. Note that this was merely a reporting issue on tensorboard, training was not impacted nor was evaluation (when using `--eval`).